pyproject.toml
[project]
name = "proXXy"
requires-python = ">=3.13"
license = { text = "GPL-3.0-only" }
authors = [{ name = "TheRealMkadmi" }]
version = "1.0.0"

dependencies = [
    "yaspin>=3.1.0,<4.0.0",
    "scrapy>=2.12.0,<3.0.0",
    "tqdm>=4.67.1,<5.0.0",
    "requests>=2.31.0,<3.0.0",
	"urllib3>=2.2.0",
	"loguru>=0.7.2",
    "proxy.py>=2.4.4,<3.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"


[tool.uv]
package = false
